fix abs has no effect warning

+++ mdrv_fw/usr/app_main.h
@@ -19,8 +19,8 @@
 
 // printf float value without enable "-u _printf_float"
 // e.g.: printf("%d.%.2d\n", P_2F(2.14));
-#define P_2F(x) (int)(x), abs(((x)-(int)(x))*100)  // "%d.%.2d"
-#define P_3F(x) (int)(x), abs(((x)-(int)(x))*1000) // "%d.%.3d"
+#define P_2F(x) (int)(x), abs((int)(((x)-(int)(x))*100))  // "%d.%.2d"
+#define P_3F(x) (int)(x), abs((int)(((x)-(int)(x))*1000)) // "%d.%.3d"
 
 
 #define BL_ARGS             0x20000000 // first word
